+++ src/Extension/Core/Widget/Paragraph/Wrap.php
@@ -4,14 +4,10 @@ declare(strict_types=1);
 
 namespace PhpTui\Tui\Extension\Core\Widget\Paragraph;
 
-final class Wrap
+enum Wrap
 {
-    private function __construct(public bool $trim)
-    {
-    }
-
-    public static function trimmed(): self
-    {
-        return new self(trim: true);
-    }
+    case None;
+    case Word;
+    case WordTrimmed;
+    case Character;
 }